UIContainer: - added a vertical safety buffer to make sure text is drawn during sprite scrolling
UIElements:
- introduced a method to check is elemnt is visible when sprite is scrolled

CAUTION:
TFT_eSPI library needs to be adjusted for this to work correctly. The method TFT_eSPI.drawString() will reset out of display/sprite coorinated to 0 or the max size:
// Check coordinates are OK, adjust if not
if (poX < 0) poX = 0;
if (poX + cwidth > width())   poX = width() - cwidth;
if (poY < 0) poY = 0;
if (poY + cheight - baseline > height()) poY = height() - cheight;

Those checks need to be commented out in order to draw strings on the outside of the valid area.